+++ lib/pvectl/plugin_loader.rb
@@ -44,6 +44,7 @@ module Pvectl
       Commands::RollbackSnapshot,
       Commands::RestoreBackup,
       Commands::Console,
+      Commands::SendkeyVm,
       Commands::Pull,
       Commands::Push,
       Commands::Service,